move hammer handlers out
